+++ romanticApi/src/main/java/cn/com/cgh/romantic/em/ResourceStatus.java
@@ -13,12 +13,12 @@ public enum ResourceStatus implements DisplayedEnum {
     /**
      * 菜单
      */
-    MENU(0, "菜单"),
+    MENU(0L, "菜单"),
     /**
      * 按钮
      */
-    BUTTON(1, "按钮");
+    BUTTON(1L, "按钮");
 
-    private final Integer value;
+    private final Long value;
     private final String label;
 }
