+++ romanticApi/src/main/java/cn/com/cgh/romantic/em/YesNoStatus.java
@@ -3,18 +3,22 @@ package cn.com.cgh.romantic.em;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+/**
+ * @author cgh
+ */
+
 @Getter
 @AllArgsConstructor
 public enum YesNoStatus implements DisplayedEnum {
     /**
      * 是
      */
-    YES(1, "是"),
+    YES(1L, "是"),
     /**
      * 否
      */
-    NO(0, "否");
+    NO(0L, "否");
 
-    private final Integer value;
+    private final Long value;
     private final String label;
 }
